Refactor/1176 vector move colormappicker (#1178)
