+++ sites/geohub/package.json
@@ -62,6 +62,7 @@
     "@watergis/legend-symbol": "^0.2.3",
     "@watergis/maplibre-gl-export": "^1.3.8",
     "bignumber.js": "^9.1.1",
+    "bulma": "^0.9.4",
     "bulma-divider": "^0.2.0",
     "bulma-o-steps": "^1.1.0",
     "bulma-slider": "^2.0.5",
